Improved the speed and effectiveness of computing the image histograms on realistic data. Fixed bug where progress dialog would close when loading imagery before the histogram creationg progress dialog would open. Forced loaded imagery to cast to float32. All image processing algorithms assume data are floating point values.

+++ vista/widgets/core/data/data_loader.py
@@ -360,7 +360,7 @@ class DataLoaderThread(QThread):
         num_images = images_dataset.shape[0]
 
         # Pre-allocate the output array
-        images = np.empty(images_dataset.shape, dtype=images_dataset.dtype)
+        images = np.empty(images_dataset.shape, dtype=np.float32)
 
         # For small datasets, read all at once
         if num_images < 10:
